requirements.txt
aws-cdk-lib==2.185.0
constructs>=10.0.0,<11.0.0
python-dotenv>=1.0.0
